+++ crates/symbolicator-service/src/services/caches/mod.rs
@@ -1,0 +1,3 @@
+//! The various caches used by the core Symbolication Service are placed here.
+
+pub mod versions;